requirements.txt
django==2.2.19
djangorestframework==3.12.4
django-filter==21.1